backend_heroku/requirements.txt
Flask==0.10.1
numpy==1.9.2
scipy==0.18.1
six==1.9.0
h5py==2.6.0
cython==0.23.4
chainer==1.14.0
Pillow==2.3.0
psycopg2==2.6.2
scikit-image==0.11.3
matplotlib==1.3.1

